refactor/project: re-org the folder structure to comply with rust conventions
