+++ Cargo.toml
@@ -16,7 +16,6 @@ categories = ["parsing", "development-tools"]
 
 [lib]
 name = "rpm"
-path = "src/rpm.rs"
 
 [dependencies]
 nom = "5.1"
