+++ src/lib.rs
@@ -1,0 +1,16 @@
+mod errors;
+pub use crate::errors::*;
+
+pub mod crypto;
+
+pub(crate) mod constants;
+pub use crate::constants::*;
+
+mod rpm;
+pub use crate::rpm::*;
+
+#[cfg(test)]
+mod tests;
+
+#[cfg(all(test, feature = "test-with-podman"))]
+mod compat_tests;
