+++ src/rpm/headers/mod.rs
@@ -1,0 +1,13 @@
+mod header;
+mod lead;
+mod types;
+
+pub use header::*;
+pub use lead::*;
+pub use types::*;
+
+#[cfg(feature = "signing-meta")]
+mod signature_builder;
+
+#[cfg(feature = "signing-meta")]
+pub use signature_builder::*;
