+++ src/rpm/headers/signature_builder.rs
@@ -2,6 +2,8 @@
 
 use super::*;
 
+use super::IndexEntry;
+use crate::constants::*;
 use std::default::Default;
 
 /// A marker trait for builder stages
@@ -133,16 +135,16 @@ mod test {
             .build(32i32);
 
         assert!(header
-            .find_entry_or_err(&constants::IndexSignatureTag::RPMSIGTAG_RSA)
+            .find_entry_or_err(&IndexSignatureTag::RPMSIGTAG_RSA)
             .is_ok());
         assert!(header
-            .find_entry_or_err(&constants::IndexSignatureTag::RPMSIGTAG_PGP)
+            .find_entry_or_err(&IndexSignatureTag::RPMSIGTAG_PGP)
             .is_ok());
         assert!(header
-            .find_entry_or_err(&constants::IndexSignatureTag::RPMSIGTAG_MD5)
+            .find_entry_or_err(&IndexSignatureTag::RPMSIGTAG_MD5)
             .is_ok());
         assert!(header
-            .find_entry_or_err(&constants::IndexSignatureTag::RPMSIGTAG_SHA1)
+            .find_entry_or_err(&IndexSignatureTag::RPMSIGTAG_SHA1)
             .is_ok());
     }
 }
