+++ src/rpm/mod.rs
@@ -1,0 +1,15 @@
+mod builder;
+mod compressor;
+mod headers;
+mod package;
+
+#[cfg(feature = "signing-meta")]
+pub mod signature;
+
+pub use headers::*;
+
+pub use compressor::*;
+
+pub use package::*;
+
+pub use builder::*;
