+++ src/rpm/signature.rs
@@ -1,7 +1,6 @@
+// re-export
 #[cfg(feature = "signing-pgp")]
 pub use crate::crypto::{
     self,
     pgp::{Signer as SignerPGP, Verifier as VerifierPGP},
 };
-
-// export more as they are created
